Migrate to C++20 and improve memory handling
Updated project files to use C++20 standard. Replaced alloca with _malloca/_freea for safer stack allocations and added null checks with std::bad_alloc throws. Improved Direct3D12 resource handling, fixed minor math and API usage issues, and performed code cleanup for initialization and type safety.

+++ Engine/Graphics/Direct3D12/Direct3D12Surface.h
@@ -64,7 +64,7 @@ namespace lightning::graphics::direct3d12 {
 				u32 height{ 0 };
 				ID3D12Fence* fence{ nullptr };
 				u32 fence_value{ 0 };
-				u32 total_bytes{ 0 };
+				u64 total_bytes{ 0 };
 			};
 
 			#if USE_STL_VECTOR
